+++ primo-sito-web-PURE/src/main/java/it/polimi/tiw/Bean/ProdottoDiUnFornitore.java
@@ -1,5 +1,5 @@
 package it.polimi.tiw.Bean;
 
-public record ProdottoDiUnFornitore(int idProdotto, int idFornitore, float prezzo, float sconto) {
+public record ProdottoDiUnFornitore(int idProdotto, int idFornitore, double prezzo, double sconto) {
 	
 }
